Fix my mistake with Bit reversal

+++ include/MultiFFT.h
@@ -21,7 +21,7 @@ namespace MultiFFT {
 	void ThirdIttFFTRC(std::vector<std::complex<double>>& fbins);
 	std::vector<std::complex<double>> ConvertSignalToComplex(const std::vector<double>& input);
 	unsigned int RoundDownPowerOfTwo(unsigned int n);
-	constexpr unsigned long long BitReverse(unsigned long long b, unsigned long long head);
+	constexpr unsigned int BitReverse(unsigned int b, unsigned int head);
 	Signal<double> HannWindowing(const Signal<double>& sample);
 	void PrintMessage(const std::string&);
 	bool IsPowerOfTwo(int x);
